Fix casting a pointer to an integer safely.
Only casting from an integer (any size) to a pointer using reinterpret_cast<>
is well-defined. Casting a pointer to an integer needs a large enough integer.

Bug 29894961

Reviewed-on: https://swiftshader-review.googlesource.com/5750
Tested-by: Nicolas Capens <capn@google.com>

+++ src/OpenGL/libEGL/libEGL.cpp
@@ -989,7 +989,7 @@ EGLImageKHR CreateImageKHR(EGLDisplay dpy, EGLContext ctx, EGLenum target, EGLCl
 		}
 	#endif
 
-	GLuint name = reinterpret_cast<GLuint>(buffer);
+	GLuint name = static_cast<GLuint>(reinterpret_cast<uintptr_t>(buffer));
 
 	if(name == 0)
 	{
